Move LevelReader into GameScreen
